st - Use require_relative to require the inheritable_fixtures' instead of adding inheritable_fixtures/lib to the load path (which is what gets generated by "bundle gem inheritable_fixtures").
The reason is that doing the latter causes this gem's lib dir to be after
its dependencies on the load path (in the bundler context of this
project) as opposed to being before them.
That is a problem for testing because we want to test the test_unit
model generator override in inheritable_fixtures/lib that overrides the
one in railties/lib, so inheritable_fixtures/lib has to be ahead of
railties/lib on the load path. This is only a test issue for this
project, does not have impact of the usage of this gem.